Rework resource pack building
- New pack task system
- New font abstraction layer
- Improved char size calculation
  - char sizes are stored as floats now

+++ nova/src/main/kotlin/xyz/xenondevs/nova/ui/overlay/actionbar/ActionbarOverlay.kt
@@ -13,7 +13,7 @@ interface ActionbarOverlay {
     /**
      * The width of the [component] in pixels.
      */
-    fun getWidth(locale: String): Int =
+    fun getWidth(locale: String): Float =
         CharSizes.calculateComponentWidth(component, locale)
     
 }
